Add authentication and authorization in each api

+++ src/main/java/com/gharelu/customer_service/controller/CustomerController.java
@@ -2,6 +2,11 @@ package com.gharelu.customer_service.controller;
 
 import com.gharelu.customer_service.model.Customer;
 import com.gharelu.customer_service.service.CustomerService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,46 +14,79 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/customer")
+@RequiredArgsConstructor
 public class CustomerController {
 
     private final CustomerService service;
 
-    public CustomerController(CustomerService service) {
-        this.service = service;
+    @Autowired
+    @Qualifier("authServiceWebClient")
+    private final WebClient webClient;
+    private String isTokenValid(String authHeader) {
+        String authResponse = webClient.get()
+                .header("Authorization", authHeader)
+                .retrieve()
+                .bodyToMono(String.class)
+                .block();
+        return authResponse;
     }
 
     @PostMapping
-    public ResponseEntity<Customer> createCustomer(@RequestBody Customer customer) {
-        return ResponseEntity.ok(service.createCustomer(customer));
+    public ResponseEntity<Customer> createCustomer(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                                                   @RequestBody Customer customer) {
+        String authResponse = isTokenValid(authHeader);
+        if(authResponse.equals("Valid"))
+            return ResponseEntity.ok(service.createCustomer(customer));
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping
-    public ResponseEntity<List<Customer>> getAllCustomers() {
-        return ResponseEntity.ok(service.getAllCustomers());
+    public ResponseEntity<List<Customer>> getAllCustomers(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader) {
+        String authResponse = isTokenValid(authHeader);
+        if(authResponse.equals("Valid"))
+            return ResponseEntity.ok(service.getAllCustomers());
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Customer> getCustomer(@PathVariable Long id) {
-        return service.getCustomerById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<Customer> getCustomer(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                                                @PathVariable Long id) {
+        String authResponse = isTokenValid(authHeader);
+        if(authResponse.equals("Valid"))
+            return service.getCustomerById(id)
+                    .map(ResponseEntity::ok)
+                    .orElse(ResponseEntity.notFound().build());
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Customer> updateCustomer(@PathVariable Long id, @RequestBody Customer updated) {
-        return ResponseEntity.ok(service.updateCustomer(id, updated));
+    public ResponseEntity<Customer> updateCustomer(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                                                   @PathVariable Long id, @RequestBody Customer updated) {
+        String authResponse = isTokenValid(authHeader);
+        if(authResponse.equals("Valid"))
+            return ResponseEntity.ok(service.updateCustomer(id, updated));
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCustomer(@PathVariable Long id) {
-        service.deleteCustomer(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> deleteCustomer(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                                               @PathVariable Long id) {
+        String authResponse = isTokenValid(authHeader);
+        if(authResponse.equals("Valid")) {
+            service.deleteCustomer(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
     }
 }
